+++ src/main/java/insane96mcp/enhancedai/modules/zombie/ZombieModule.java
@@ -1,6 +1,7 @@
 package insane96mcp.enhancedai.modules.zombie;
 
 import insane96mcp.enhancedai.modules.zombie.feature.DiggerZombieFeature;
+import insane96mcp.enhancedai.modules.zombie.feature.PearlerZombieFeature;
 import insane96mcp.enhancedai.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -9,17 +10,20 @@ import insane96mcp.insanelib.base.Module;
 public class ZombieModule extends Module {
 
 	public DiggerZombieFeature diggerZombie;
+	public PearlerZombieFeature pearlerZombie;
 
 	public ZombieModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
-		diggerZombie = new DiggerZombieFeature(this);
+		this.diggerZombie = new DiggerZombieFeature(this);
+		this.pearlerZombie = new PearlerZombieFeature(this);
 		Config.builder.pop();
 	}
 
 	@Override
 	public void loadConfig() {
 		super.loadConfig();
-		diggerZombie.loadConfig();
+		this.diggerZombie.loadConfig();
+		this.pearlerZombie.loadConfig();
 	}
 }