+++ src/main/java/insane96mcp/enhancedai/modules/zombie/ai/AIZombiePearler.java
@@ -1,6 +1,5 @@
 package insane96mcp.enhancedai.modules.zombie.ai;
 
-import net.minecraft.command.arguments.EntityAnchorArgument;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.item.EnderPearlEntity;
@@ -12,11 +11,15 @@ import net.minecraft.item.Items;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.vector.Vector3d;
 
 public class AIZombiePearler extends Goal {
 
 	private final ZombieEntity pearler;
 	private PlayerEntity targetPlayer;
+	private int cooldown = 40;
+
+	EnderPearlEntity enderPearlEntity;
 
 	public AIZombiePearler(ZombieEntity pearler){
 		this.pearler = pearler;
@@ -27,34 +30,46 @@ public class AIZombiePearler extends Goal {
 		if (!(target instanceof PlayerEntity))
 			return false;
 
-		if (this.pearler.getDistanceSq(target) < 16d * 16d)
+		if (this.pearler.getDistanceSq(target) < 12d * 12d)
+			return false;
+
+		if (--this.cooldown > 0)
+			return false;
+
+		if (!this.pearler.isOnGround())
 			return false;
 
 		return this.pearler.getHeldItemMainhand().getItem() == Items.ENDER_PEARL || this.pearler.getHeldItemOffhand().getItem() == Items.ENDER_PEARL;
 	}
 
 	public boolean shouldContinueExecuting() {
-		return false;
+		return this.enderPearlEntity != null && this.enderPearlEntity.isAlive();
 	}
 
 	public void startExecuting() {
 		this.targetPlayer = (PlayerEntity) this.pearler.getAttackTarget();
 		EquipmentSlotType slot = this.pearler.getHeldItemMainhand().getItem() == Items.ENDER_PEARL ? EquipmentSlotType.MAINHAND : EquipmentSlotType.OFFHAND;
-		this.pearler.world.playSound(null, this.pearler.getPosX(), this.pearler.getPosY(), this.pearler.getPosZ(), SoundEvents.ENTITY_ENDER_PEARL_THROW, SoundCategory.NEUTRAL, 0.5F, 0.4F / (this.pearler.world.rand.nextFloat() * 0.4F + 0.8F));
+		this.pearler.world.playSound(null, this.pearler.getPosX(), this.pearler.getPosY(), this.pearler.getPosZ(), SoundEvents.ENTITY_ENDER_PEARL_THROW, SoundCategory.NEUTRAL, 1F, 0.4F / (this.pearler.world.rand.nextFloat() * 0.4F + 0.8F));
 		ItemStack stack = this.pearler.getItemStackFromSlot(slot);
-		EnderPearlEntity enderpearlentity = new EnderPearlEntity(this.pearler.world, this.pearler);
-		enderpearlentity.setItem(stack);
-		this.pearler.lookAt(EntityAnchorArgument.Type.EYES, this.targetPlayer.getEyePosition(1f));
-		double d0 = this.targetPlayer.getPosX() - this.pearler.getPosX();
-		double d1 = this.targetPlayer.getPosYHeight(0.3333333333333333D) - this.pearler.getPosYHeight(0.3333333333333333D);
-		double d2 = this.targetPlayer.getPosZ() - this.pearler.getPosZ();
-		double d3 = (double) MathHelper.sqrt(d0 * d0 + d2 * d2);
-		enderpearlentity.setDirectionAndMovement(this.pearler, this.pearler.rotationPitch, this.pearler.rotationYaw, 0.0F, 1.5F, 0);
-		this.pearler.world.addEntity(enderpearlentity);
+		this.enderPearlEntity = new EnderPearlEntity(this.pearler.world, this.pearler);
+		enderPearlEntity.setPosition(this.pearler.getEyePosition(1f).x, this.pearler.getEyePosition(1f).y, this.pearler.getEyePosition(1f).z);
+		enderPearlEntity.setItem(stack);
+		Vector3d vector3d = this.pearler.getEyePosition(1f);
+		double d0 = this.targetPlayer.getPosX() - vector3d.x;
+		double d1 = this.targetPlayer.getEyePosition(1f).y - vector3d.y;
+		double d2 = this.targetPlayer.getPosZ() - vector3d.z;
+		double d3 = MathHelper.sqrt(d0 * d0 + d2 * d2);
+		double pitch = MathHelper.wrapDegrees((float)(-(MathHelper.atan2(d1, d3) * (double)(180F / (float)Math.PI))));
+		double yaw = MathHelper.wrapDegrees((float)(MathHelper.atan2(d2, d0) * (double)(180F / (float)Math.PI)) - 90.0F);
+		enderPearlEntity.setDirectionAndMovement(this.pearler, (float) (pitch - 3f - d1), (float) (yaw), 0.0F, 1.5F, 5);
+		this.pearler.world.addEntity(enderPearlEntity);
 		stack.shrink(1);
+		this.cooldown = 60;
 	}
 
 	public void resetTask() {
 		this.targetPlayer = null;
+		this.enderPearlEntity = null;
+		this.pearler.getNavigator().clearPath();
 	}
 }